feat: extract team builder logic into reusable composable
- Create comprehensive useTeamBuilder.ts composable (360 lines)
- Extract all business logic from TeamBuilderForm.vue and index.vue
- Centralise data fetching, team management, and form validation
- Implement structured loading states (fetchingTeam, submittingForm)
- Add standardised error handling with consistent user feedback
- Simplify page component from 110 to 30 lines
- Simplify form component from 200+ to 80 lines
- Maintain TypeScript strict mode compliance
- Update task documentation with completion status
